Added new features: Detect shipping price
No longer uses the webscraper.py module, since we need to parse ebay all in the same requests Session. I suppose we could work that in, but this is easier. Description does not work for anybody who has a customized description, so it is a work in progress. Adding images soon!